Refactor of Model objects. Still needs fixing
